backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic[email]==2.5.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
python-dotenv==1.0.0
scikit-learn==1.4.0
pandas==2.2.0
joblib==1.3.2
numpy==1.26.3
reportlab==4.0.9
qrcode[pil]==7.4.2
google-generativeai==0.7.2
